% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildURL.R
\name{buildURL}
\alias{buildURL}
\title{Helper function to build a URL matching the VC API format}
\usage{
buildURL(apiKey = NULL, table = NULL, ...)
}
\arguments{
\item{apiKey}{User-specific API Key provided by Visual Crossing Weather}

\item{table}{Name of the table the data will come from}

\item{...}{Named arguments specific to the selected table}
}
\value{
String. The URL required to access the data at VC Weather
}
\description{
Helper function to build a URL matching the VC API format
}
