% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{northeast}
\alias{northeast}
\title{Gives 4 locations in the 'Northeast' US Climate Region}
\usage{
northeast()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'Northeast' US Climate Region
}
\examples{
northeast()
}
