% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{northRockies}
\alias{northRockies}
\title{Gives 4 locations in the 'Northern Rockies and Plains' US Climate Region}
\usage{
northRockies()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'Northern Rockies and Plains' US Climate Region
}
\examples{
northRockies()
}
