% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{northwest}
\alias{northwest}
\title{Gives 4 locations in the 'Northwest' US Climate Region}
\usage{
northwest()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'Northwest' US Climate Region
}
\examples{
northwest()
}
