% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{ohioValley}
\alias{ohioValley}
\title{Gives 4 locations in the 'Ohio Valley' US Climate Region}
\usage{
ohioValley()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'Ohio Valley' US Climate Region
}
\examples{
ohioValley()
}
