% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{south}
\alias{south}
\title{Gives 4 locations in the 'South' US Climate Region}
\usage{
south()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'South' US Climate Region
}
\examples{
south()
}
