% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{southeast}
\alias{southeast}
\title{Gives 4 locations in the 'Southeast' US Climate Region}
\usage{
southeast()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'Southeast' US Climate Region
}
\examples{
southeast()
}
