% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{southwest}
\alias{southwest}
\title{Gives 4 locations in the 'Southwest' US Climate Region}
\usage{
southwest()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'Southwest' US Climate Region
}
\examples{
southwest()
}
