% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{upperMidwest}
\alias{upperMidwest}
\title{Gives 4 locations in the 'Upper Midwest' US Climate Region}
\usage{
upperMidwest()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'Upper Midwest' US Climate Region
}
\examples{
upperMidwest()
}
