% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{USregions}
\alias{USregions}
\alias{northwest}
\alias{northRockies}
\alias{upperMidwest}
\alias{northeast}
\alias{west}
\alias{southwest}
\alias{south}
\alias{ohioValley}
\alias{southeast}
\title{Selection of cities in Continental US Climate Regions}
\usage{
northwest()

northRockies()

upperMidwest()

northeast()

west()

southwest()

south()

ohioValley()

southeast()
}
\value{
A character vector in the format Cityname, ST.
}
\description{
These functions return (somewhat haphazardly chosen) cities in Continental US climate
regions as defined by Karl and Koss (1984).
}
\examples{
northwest()
northRockies()
upperMidwest()
northeast()
west()
southwest()
south()
ohioValley()
southeast()
}
