% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{vc.history}
\alias{vc.history}
\title{Get data from the 'history' table on VC Weather}
\usage{
vc.history(
  apiKey,
  locations = NULL,
  startDateTime = NULL,
  endDateTime = NULL,
  period = NULL,
  aggregateHours = "24",
  aggregateMinutes = NULL,
  combinationMethod = "aggregate",
  unitGroup = "us",
  outputDateTimeFormat = NULL,
  dayEndTime = "0:0:00",
  dayStartTime = dayEndTime,
  timezone = NULL,
  collectStationContribution = "false",
  includeAstronomy = "false",
  extendedStats = "false",
  maxDistance = "50000",
  shortColumnNames = "false",
  contentType = "csv",
  locationMode = "single",
  iconSet = "icons1",
  lang = "en"
)
}
\arguments{
\item{apiKey}{User-specific API Key provided by Visual Crossing Weather}

\item{locations}{One or more address, partial address or latitude, longitude values for the required locations . Addresses can be specified as full addresses. The system will also attempt to match partial addresses such as city, state, zip code, postal code and other common formats.

When specifying a point based on longitude, latitude, the format must be specified as latitude,longitude where both latitude and longitude are in decimal degrees. latitude should run from -90 to 90 and longitude from -180 to 180 (with 0 being at the prime meridian through London, UK).

Data for multiple locations can be requested in a single request by concatenating multiple locations using the pipe (|) character.}

\item{startDateTime}{The date time for the start of the data request using the time zone of the location. In the ISO format: yyyy-MM-ddTHH:mm:ss. Hours should be specified in 24 hour clock format.}

\item{endDateTime}{The date time for the end of the data request using the time zone of the location. In the ISO format: yyyy-MM-ddTHH:mm:ss. Hours should be specified in 24 hour clock format.}

\item{period}{Named data time range used instead of the startDateTime and endDateTime parameters. Available period values are “today”,”yesterday”,tomorrow”,”yeartodate”,”last30days”,”lastyear”,”last24hours”,”next30days”,”next7days” and “nextyear”.}

\item{aggregateHours}{The interval between weather history data in the output. 1 represent hourly records, 24 represents a daily forecast. As the source data is recorded at the hourly level, 24 hour records are aggregated to indicate the predominant weather conditions during that time period. See also ‘aggregateMinutes’. Supported values 1 or 24, default is 24.}

\item{aggregateMinutes}{The interval between weather history data in the output in minutes. This requires a plan that includes access to sub-hourly data. This parameter is used instead of ‘aggregateHours’. Use a factor of 60, such as 5,10,15 or 30.}

\item{combinationMethod}{Describes how multiple individual weather station observations should be combined. Options are aggregate, best and list.}

\item{unitGroup}{The system of units used for the output data. Supported values are us,uk,metric,base.}

\item{outputDateTimeFormat}{specifies the date and time format used when returning the output.}

\item{dayEndTime}{when present and not set to the same as the dayEndTime, filters the output to records that between the specified day times. When specified, the result can have no more than 30 days.}

\item{dayStartTime}{When present and not set to the same as the dayEndTime, Filters the output to records that between the specified day times. This is useful for setting filters for business hours. Format h:m:ss (eg 9:00:00 would be 9am). When specified, the result can have no more than 30 days.}

\item{timezone}{specifies the timezone of the input and result dates and times. When not specified, all date times are considered local times. if you would like to specify that all dates are entered as UTC dates and times, use timezone=Z parameter.}

\item{collectStationContribution}{Whether to include a column describing the weather stations that were used for a particular output record. This can vary at the individual record level if weather stations do not transmit a full record and other stations are used as back ups.}

\item{includeAstronomy}{Retrieve astronomical data such as moon phase, sunrise and sunset times. Set to true to enable.}

\item{extendedStats}{Retrieve additional statistical values in the weather request including mean, maximum and minimum of the wind speed, wind chill and heat index. Set to true to enable.}

\item{maxDistance}{The maximum distance in meters used to search for local weather stations (default 50000m). This setting is combined with the maxStations parameter to find local weather stations.}

\item{shortColumnNames}{When false, the returned dataset includes descriptive column names. When true, returns shorter, abbreviated column names with only alphanumeric characters. The short names are useful for programmatic use of the data and database data import.}

\item{contentType}{Selected between CSV or json output format. Supported values are ‘csv’ or ‘json’}

\item{locationMode}{Requests the structure that will be used for the output locations in the JSON output format. Supported values: single,array, or lookup. See the VC JSON documentation for more information on how this parameters affects the output JSON structure.}

\item{iconSet}{returns fixed set of of the icons names based on the weather conditions. Currently supported value is ‘ icons1’}

\item{lang}{Sets the language of the translatable parts of the output such as the conditions field. Available languages are en, de, fr and es. In addition passing in ‘id’ will result in the raw descriptor IDs.}
}
\value{
data.frame with output of VC Weather call
}
\description{
This function guides construction of a request to the 'history' table on the VC Weather API. The parameter descriptions are quoted directly from the VC weather-api documentation and should be entered as character strings.
}
\examples{
\dontrun{tst <- vc.history('APIKEY', locations = 'Rochester, NY',
                            startDateTime = '1993-12-20T00:00:00',
                            endDateTime = '1993-12-22T00:00:00')}
}
