% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCrossingAPI-package.R
\docType{package}
\name{visCrossingAPI-package}
\alias{visCrossingAPI}
\alias{visCrossingAPI-package}
\title{visCrossingAPI: Interface to Visual Crossing Weather API}
\description{
Provides an easy to use R interface to the Visual Crossing Weather API. The data is owned and provided by Visual Crossing Weather, use of this package requires an API key provided by Visual Crossing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeDuttweiler/visCrossingAPI}
  \item Report bugs at \url{https://github.com/LukeDuttweiler/visCrossingAPI/issues}
}

}
\author{
\strong{Maintainer}: Luke Duttweiler \email{lduttweiler@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-0467-995X}{ORCID}) [copyright holder]

}
\keyword{internal}
