% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{west}
\alias{west}
\title{Gives 4 locations in the 'West' US Climate Region}
\usage{
west()
}
\value{
Vector of City, State Names
}
\description{
Gives 4 locations in the 'West' US Climate Region
}
\examples{
west()
}
